Added memdump. Causes crash. o3ds, n3ds dumps partial
I added a memory dump to the function and it dumps memory to /smilehook/memdumps/memdump.bin but it causes a crash when the defined range in ram is too large? A read or write access abort happens.
Don't know why. It looks like in this I'm reading from process ram or executable but not actually userland ram where SmileBasic Stores Variables. How to reach it without a crash?